# src/models/CategoryModel.ts
import { DataTypes, Model, Sequelize } from "sequelize";
import sequelize from "../config/database"; // Adjust the path as necessary

class CategoryModel extends Model{
    id: number | undefined;
    name: string | undefined;
}

CategoryModel.init(
    {
        id: {
            type: DataTypes.INTEGER,
            autoIncrement: true,
            primaryKey: true
        },
        name: {
            type: DataTypes.STRING,
            allowNull: false
        }
    },
    {
        sequelize,
        modelName: "CategoryModel",
        tableName: "category"
    }
)

export default CategoryModel;